Rest APIEndpoint3: Refactor service class
